0712 - 소수점 출력 수정

+++ project-1/src/main/java/com/kosmo/springapp/model/TotalReviewDTO.java
@@ -20,7 +20,7 @@ import lombok.Setter;
 @Component
 @Alias("totalReviewDto")
 public class TotalReviewDTO {
-	int starScoreTotal;
+	float starScoreTotal;
 	Map starScore;
 	List<Map<String,Integer>> effectList;
 	List<Map<String,Integer>> noEffectList;
